refactor: move product methods to separate class
