+++ src/pages/product/ui/Product.tsx
@@ -11,7 +11,7 @@ export function Product() {
   const [product, setProduct] = useState<ProductData>(emptyProduct);
 
   async function getProduct() {
-    const product = await client.getProduct(id);
+    const product = await client.productApi.getProduct(id);
     setProduct(product);
   }
