+++ src/shared/api/clientApi/ClientApi.ts
@@ -9,19 +9,11 @@ import {
 } from './CreateApiRoots';
 import {
   AccountAddress,
-  QueryMode,
-  SortingTypes,
   type loginDTO,
-  type MainCategory,
-  type PriceRange,
-  type ProductData,
   type ProfileData,
-  type RangeObject,
   type singUpDTO,
 } from './types';
-import { parseProduct } from './parsers/parseProduct';
-import { DefaultProfileData, emptyProduct, productPerPage } from './constants';
-import { parseCategories } from './parsers/parseCategories';
+import { DefaultProfileData } from './constants';
 import { parseProfileData } from './parsers/parseProfileData';
 import { createSignUpBody } from './utils/createSignUpBody';
 import {
@@ -32,34 +24,21 @@ import {
 import { countryCodes } from './utils/CountryCodes';
 import { parseCartData } from './parsers/parseCartData';
 import { CartApi } from './subclasses/CartApi';
+import { ProductApi } from './subclasses/ProductApi';
 
 class ClientApi {
   public isLogin: boolean;
-  public categories: MainCategory[];
-  public currentCategoryId: string;
   public profileData: ProfileData;
-  public priceRange: PriceRange;
-  public minRating: string;
-  public sortingType: string;
-  public searchText: string;
-  public queryMode: string;
-  public pageCount: number;
+  public productApi: ProductApi;
   public cartApi: CartApi;
   private apiRoot: ByProjectKeyRequestBuilder;
 
   constructor() {
     this.isLogin = false;
     this.apiRoot = CreateAnonymousApiRoot();
-    this.currentCategoryId = '';
-    this.categories = [];
     this.profileData = DefaultProfileData;
-    this.priceRange = { min: 0, max: 100 };
-    this.minRating = '1';
-    this.sortingType = SortingTypes.DEFAULT;
-    this.searchText = '';
-    this.queryMode = QueryMode.FILTER;
-    this.pageCount = 1;
 
+    this.productApi = new ProductApi(this.apiRoot);
     this.cartApi = new CartApi(this.apiRoot);
   }
 
@@ -241,136 +220,6 @@ class ClientApi {
     this.profileData = parseProfileData(results.body);
     this.profileData.version = results.body.version;
     return this.profileData.accountAddresses;
-  }
-
-  public getCategoryName(id: string): string {
-    const category = this.categories.find((item) => item.id === id);
-    if (category) return category.name;
-    return '';
-  }
-
-  public async getProducts(pageIndex?: number): Promise<ProductData[]> {
-    try {
-      this.queryMode = QueryMode.FILTER;
-
-      const response = await this.apiRoot
-        .productProjections()
-        .search()
-        .get({
-          queryArgs: {
-            filter: [
-              `categories.id:"${this.currentCategoryId}"`,
-              `variants.price.centAmount: range (${this.priceRange.min} to ${this.priceRange.max + 1})`,
-              `variants.attributes.rating: range (${client.minRating} to 6)`,
-            ],
-            limit: productPerPage,
-            offset: pageIndex ? (pageIndex - 1) * productPerPage : 0,
-          },
-        })
-        .execute();
-
-      const countProduct = response.body.total || 0;
-      this.pageCount = Math.ceil(countProduct / productPerPage);
-      if (this.pageCount === 0) this.pageCount = 1;
-
-      const results = response.body.results;
-      const products: ProductData[] = results.map((result) =>
-        parseProduct(result)
-      );
-      return products;
-    } catch {
-      return [];
-    }
-  }
-
-  public async searchProducts(pageIndex?: number): Promise<ProductData[]> {
-    try {
-      this.queryMode = QueryMode.SEARCH;
-
-      const response = await this.apiRoot
-        .productProjections()
-        .search()
-        .get({
-          queryArgs: {
-            'text.en-US': `${client.searchText}`,
-            fuzzy: true,
-            sort:
-              client.sortingType === SortingTypes.DEFAULT
-                ? undefined
-                : client.sortingType,
-            limit: productPerPage,
-            offset: pageIndex ? (pageIndex - 1) * productPerPage : 0,
-          },
-        })
-        .execute();
-
-      const countProduct = response.body.results.length;
-      this.pageCount = Math.ceil(countProduct / productPerPage);
-      if (this.pageCount === 0) this.pageCount = 1;
-
-      const results = response.body.results;
-      const products: ProductData[] = results.map((result) =>
-        parseProduct(result)
-      );
-      return products;
-    } catch {
-      return [];
-    }
-  }
-
-  public async getProduct(id: string | undefined): Promise<ProductData> {
-    try {
-      if (id === undefined) return emptyProduct;
-      const response = await this.apiRoot
-        .productProjections()
-        .search()
-        .get({
-          queryArgs: {
-            filter: `id:"${id}"`,
-          },
-        })
-        .execute();
-      const result = response.body.results[0];
-      return parseProduct(result);
-    } catch {
-      return emptyProduct;
-    }
-  }
-
-  public async getMainCategories(): Promise<void> {
-    try {
-      const response = await this.apiRoot.categories().get().execute();
-      this.categories = parseCategories(response.body.results);
-      this.currentCategoryId = this.categories[0].id;
-    } catch {
-      this.categories = [];
-    }
-  }
-
-  public async getMinMaxPrice(): Promise<PriceRange> {
-    const response = await this.apiRoot
-      .productProjections()
-      .search()
-      .get({
-        queryArgs: {
-          facet: 'variants.price.centAmount: range(0 to *)',
-        },
-      })
-      .execute();
-    if (!response.body.facets) return { min: 0, max: 100 };
-    const range = response.body.facets['variants.price.centAmount'] as unknown;
-    if (range && typeof range === 'object') {
-      const rangeObject: Partial<RangeObject> = range;
-      if (rangeObject.ranges !== undefined) {
-        const range = rangeObject.ranges[0];
-        this.priceRange = { min: range.min, max: range.max };
-        return {
-          min: Math.floor(range.min / 100),
-          max: Math.ceil(range.max / 100),
-        };
-      }
-    }
-    return { min: 0, max: 100 };
   }
 }
